DarkLens/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
google-generativeai==0.8.5
pydantic==2.11.1
python-dotenv==1.1.0
Pillow==11.2.1
reportlab==4.0.9
playwright==1.48.0
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.25.2